+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/command_generators/velocity_command_generator.py
@@ -1,0 +1,257 @@
+# Copyright (c) 2022, NVIDIA CORPORATION & AFFILIATES, ETH Zurich, and University of Toronto
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+"""Sub-module containing command generators for the velocity-based locomotion task."""
+
+import torch
+from typing import Sequence, Tuple
+
+from omni.isaac.orbit.markers import VisualizationMarkers
+from omni.isaac.orbit.markers.config import ARROW_X_MARKER_CFG
+from omni.isaac.orbit.robots.robot_base import RobotBase
+from omni.isaac.orbit.utils.math import quat_apply, quat_from_euler_xyz, wrap_to_pi
+
+from .command_generator_base import CommandGeneratorBase
+from .command_generator_cfg import NormalVelocityCommandGeneratorCfg, UniformVelocityCommandGeneratorCfg
+
+
+class UniformVelocityCommandGenerator(CommandGeneratorBase):
+    r"""Command generator that generates a velocity command in SE(2) from uniform distribution.
+
+    The command comprises of a linear velocity in x and y direction and an angular velocity around
+    the z-axis. It is given in the robot's base frame.
+
+    If the :attr:`cfg.heading_command` flag is set to True, the angular velocity is computed from the heading
+    error similar to doing a proportional control on the heading error. The target heading is sampled uniformly
+    from the provided range. Otherwise, the angular velocity is sampled uniformly from the provided range.
+
+    Mathematically, the angular velocity is computed as follows from the heading command:
+
+    .. math::
+
+        \omega_z = \frac{1}{2} \text{wrap_to_pi}(\theta_{\text{target}} - \theta_{\text{current}})
+
+    """
+
+    cfg: UniformVelocityCommandGeneratorCfg
+    """The configuration of the command generator."""
+
+    def __init__(self, cfg: UniformVelocityCommandGeneratorCfg, env: object):
+        """Initialize the command generator.
+
+        Args:
+            cfg (UniformVelocityCommandGeneratorCfg): The configuration of the command generator.
+            env (object): The environment.
+        """
+        super().__init__(cfg, env)
+        # -- robot
+        # TODO: Should we make this configurable like this?
+        self.robot: RobotBase = getattr(env, cfg.robot_attr)
+        # -- command: x vel, y vel, yaw vel, heading
+        self.vel_command_b = torch.zeros(self.num_envs, 3, device=self.device)
+        self.heading_target = torch.zeros(self.num_envs, device=self.device)
+        self.is_heading_env = torch.zeros(self.num_envs, dtype=torch.bool, device=self.device)
+        self.is_standing_env = torch.zeros_like(self.is_heading_env)
+        # -- metrics
+        self.metrics["error_vel_xy"] = torch.zeros(self.num_envs, device=self.device)
+        self.metrics["error_vel_yaw"] = torch.zeros(self.num_envs, device=self.device)
+        # -- debug vis
+        self._base_vel_goal_markers = None
+        self._base_vel_markers = None
+
+    def __str__(self) -> str:
+        """Return a string representation of the command generator."""
+        msg = "UniformVelocityCommandGenerator:\n"
+        msg += f"\tCommand dimension: {tuple(self.command.shape[1:])}\n"
+        msg += f"\tResampling time range: {self.cfg.resampling_time_range}\n"
+        msg += f"\tHeading command: {self.cfg.heading_command}\n"
+        if self.cfg.heading_command:
+            msg += f"\tHeading probability: {self.cfg.rel_heading_envs}\n"
+        msg += f"\tStanding probability: {self.cfg.rel_standing_envs}"
+        return msg
+
+    """
+    Properties
+    """
+
+    @property
+    def command(self) -> torch.Tensor:
+        """The desired base velocity command in the base frame. Shape is (num_envs, 3)."""
+        return self.vel_command_b
+
+    """
+    Operations.
+    """
+
+    def debug_vis(self):
+        if self.cfg.debug_vis:
+            # create markers if necessary
+            if self._base_vel_goal_markers is None:
+                marker_cfg = ARROW_X_MARKER_CFG
+                marker_cfg.markers["arrow"].color = (1.0, 0.0, 0.0)
+                self._base_vel_goal_markers = VisualizationMarkers("/Visuals/base_velocity_goal", marker_cfg)
+            if self._base_vel_markers is None:
+                marker_cfg = ARROW_X_MARKER_CFG
+                marker_cfg.markers["arrow"].color = (0.0, 0.0, 1.0)
+                self._base_vel_markers = VisualizationMarkers("/Visuals/base_velocity_current", marker_cfg)
+            # get marker location
+            # -- base state
+            base_pos_w = self.robot.data.root_pos_w.clone()
+            base_pos_w[:, 2] += 0.5
+            # -- command
+            vel_des_arrow_scale, vel_des_arrow_quat = self._resolve_xy_velocity_to_arrow(self.command[:, :2])
+            vel_arrow_scale, vel_arrow_quat = self._resolve_xy_velocity_to_arrow(self.robot.data.root_lin_vel_w[:, :2])
+            # -- goal
+            self._base_vel_goal_markers.visualize(base_pos_w, vel_des_arrow_quat, vel_des_arrow_scale)
+            # -- base velocity
+            self._base_vel_markers.visualize(base_pos_w, vel_arrow_quat, vel_arrow_scale)
+
+    """
+    Implementation specific functions.
+    """
+
+    def _resample_command(self, env_ids: Sequence[int]):
+        # sample velocity commands
+        r = torch.empty(len(env_ids), device=self.device)
+        # -- linear velocity - x direction
+        self.vel_command_b[env_ids, 0] = r.uniform_(*self.cfg.ranges.lin_vel_x)
+        # -- linear velocity - y direction
+        self.vel_command_b[env_ids, 1] = r.uniform_(*self.cfg.ranges.lin_vel_y)
+        # -- ang vel yaw - rotation around z
+        self.vel_command_b[env_ids, 2] = r.uniform_(*self.cfg.ranges.ang_vel_z)
+        # heading target
+        if self.cfg.heading_command:
+            self.heading_target[env_ids] = r.uniform_(*self.cfg.ranges.heading)
+            # update heading envs
+            self.is_heading_env[env_ids] = r.uniform_(0.0, 1.0) <= self.cfg.rel_heading_envs
+        # update standing envs
+        self.is_standing_env[env_ids] = r.uniform_(0.0, 1.0) <= self.cfg.rel_standing_envs
+
+    def _update_command(self):
+        """Post-processes the velocity command.
+
+        This function sets velocity command to zero for standing environments and computes angular
+        velocity from heading direction if the heading_command flag is set.
+        """
+        # Compute angular velocity from heading direction
+        if self.cfg.heading_command:
+            # resolve indices of heading envs
+            heading_env_ids = self.is_heading_env.nonzero(as_tuple=False).flatten()
+            # obtain heading direction
+            forward = quat_apply(
+                self.robot.data.root_quat_w[heading_env_ids, :], self.robot._FORWARD_VEC_B[heading_env_ids]
+            )
+            heading = torch.atan2(forward[:, 1], forward[:, 0])
+            # compute angular velocity
+            self.vel_command_b[heading_env_ids, 2] = torch.clip(
+                0.5 * wrap_to_pi(self.heading_target[heading_env_ids] - heading),
+                min=self.cfg.ranges.ang_vel_z[0],
+                max=self.cfg.ranges.ang_vel_z[1],
+            )
+        # Enforce standing (i.e., zero velocity command) for standing envs
+        # TODO: check if conversion is needed
+        standing_env_ids = self.is_standing_env.nonzero(as_tuple=False).flatten()
+        self.vel_command_b[standing_env_ids, :] = 0.0
+
+    def _update_metrics(self):
+        # time for which the command was executed
+        max_command_time = self.cfg.resampling_time_range[1]
+        # logs data
+        self.metrics["error_vel_xy"] += (
+            torch.norm(self.vel_command_b[:, :2] - self.robot.data.root_lin_vel_b[:, :2]) / max_command_time
+        )
+        self.metrics["error_vel_yaw"] += (
+            torch.abs(self.vel_command_b[:, 2] - self.robot.data.root_ang_vel_b[:, 2]) / max_command_time
+        )
+
+    """
+    Internal helpers.
+    """
+
+    def _resolve_xy_velocity_to_arrow(self, xy_velocity: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
+        """Converts the XY base velocity command to arrow direction rotation."""
+        # obtain default scale of the marker
+        default_scale = self._base_vel_goal_markers.cfg.markers["arrow"].scale
+        # arrow-scale
+        arrow_scale = torch.tensor(default_scale, device=self.device).repeat(xy_velocity.shape[0], 1)
+        arrow_scale[:, 0] *= torch.linalg.norm(xy_velocity, dim=1)
+        # arrow-direction
+        heading_angle = torch.atan2(xy_velocity[:, 1], xy_velocity[:, 0])
+        zeros = torch.zeros_like(heading_angle)
+        arrow_quat = quat_from_euler_xyz(zeros, zeros, heading_angle)
+
+        return arrow_scale, arrow_quat
+
+
+class NormalVelocityCommandGenerator(UniformVelocityCommandGenerator):
+    """Command generator that generates a velocity command in SE(2) from a normal distribution.
+
+    The command comprises of a linear velocity in x and y direction and an angular velocity around
+    the z-axis. It is given in the robot's base frame.
+
+    The command is sampled from a normal distribution with mean and standard deviation specified in
+    the configuration. With equal probability, the sign of the individual components is flipped.
+    """
+
+    cfg: NormalVelocityCommandGeneratorCfg
+    """The command generator configuration."""
+
+    def __init__(self, cfg: NormalVelocityCommandGeneratorCfg, env: object):
+        """Initializes the command generator.
+
+        Args:
+            cfg (NormalVelocityCommandGeneratorCfg): The command generator configuration.
+            env: The environment.
+        """
+        super().__init__(self, cfg, env)
+        # create buffers for zero commands envs
+        self.is_zero_vel_x_env = torch.zeros(self.num_envs, dtype=torch.bool, device=self.device)
+        self.is_zero_vel_y_env = torch.zeros_like(self.is_zero_vel_x_env)
+        self.is_zero_vel_yaw_env = torch.zeros_like(self.is_zero_vel_x_env)
+
+    def __str__(self) -> str:
+        """Return a string representation of the command generator."""
+        msg = "NormalVelocityCommandGenerator:\n"
+        msg += f"\tCommand dimension: {tuple(self.command.shape[1:])}\n"
+        msg += f"\tResampling time range: {self.cfg.resampling_time_range}\n"
+        msg += f"\tStanding probability: {self.cfg.rel_standing_envs}"
+        return msg
+
+    def _resample_command(self, env_ids):
+        # sample velocity commands
+        r = torch.empty(len(env_ids), device=self.device)
+        # -- linear velocity - x direction
+        self.vel_command_b[env_ids, 0] = r.normal_(mean=self.cfg.ranges.mean_vel[0], std=self.cfg.ranges.std_vel[0])
+        self.vel_command_b[env_ids, 0] *= torch.where(r.uniform_(0.0, 1.0) <= 0.5, 1.0, -1.0)
+        # -- linear velocity - y direction
+        self.vel_command_b[env_ids, 1] = r.normal_(mean=self.cfg.ranges.mean_vel[1], std=self.cfg.ranges.std_vel[1])
+        self.vel_command_b[env_ids, 1] *= torch.where(r.uniform_(0.0, 1.0) <= 0.5, 1.0, -1.0)
+        # -- angular velocity - yaw direction
+        self.vel_command_b[env_ids, 2] = r.normal_(mean=self.cfg.ranges.mean_vel[2], std=self.cfg.ranges.std_vel[2])
+        self.vel_command_b[env_ids, 2] *= torch.where(r.uniform_(0.0, 1.0) <= 0.5, 1.0, -1.0)
+
+        # update element wise zero velocity command
+        # TODO what is zero prob ?
+        self.is_zero_vel_x_env[env_ids] = r.uniform_(0.0, 1.0) <= self.cfg.ranges.zero_prob[0]
+        self.is_zero_vel_y_env[env_ids] = r.uniform_(0.0, 1.0) <= self.cfg.ranges.zero_prob[1]
+        self.is_zero_vel_yaw_env[env_ids] = r.uniform_(0.0, 1.0) <= self.cfg.ranges.zero_prob[2]
+
+        # update standing envs
+        self.is_standing_env[env_ids] = r.uniform_(0.0, 1.0) <= self.cfg.rel_standing_envs
+
+    def _update_command(self):
+        """Sets velocity command to zero for standing envs."""
+        # Enforce standing (i.e., zero velocity command) for standing envs
+        standing_env_ids = self.is_standing_env.nonzero(as_tuple=False).flatten()  # TODO check if conversion is needed
+        self.vel_command_b[standing_env_ids, :] = 0.0
+
+        # Enforce zero velocity for individual elements
+        # TODO: check if conversion is needed
+        zero_vel_x_env_ids = self.is_zero_vel_x_env.nonzero(as_tuple=False).flatten()
+        zero_vel_y_env_ids = self.is_zero_vel_y_env.nonzero(as_tuple=False).flatten()
+        zero_vel_yaw_env_ids = self.is_zero_vel_yaw_env.nonzero(as_tuple=False).flatten()
+        self.vel_command_b[zero_vel_x_env_ids, 0] = 0.0
+        self.vel_command_b[zero_vel_y_env_ids, 1] = 0.0
+        self.vel_command_b[zero_vel_yaw_env_ids, 2] = 0.0